fix: Fixing project structure for Dart project publishing. #1
