+++ test/dart_cid_test.dart
@@ -4,13 +4,13 @@ import 'dart:io';
 
 import 'package:test_process/test_process.dart';
 
-import 'package:dart_cid/dart_cid.dart';
+import 'package:dart_cid/src/cid.dart';
 
 void main() {
   group("Regular tests", () {
     test('creating cid with \'hello world \' with base58 encoding', () {
       String input = 'hello world';
-      final output = createCid(input, BASE.base58);
+      final output = Cid.createCid(input, BASE.base58);
 
       // https://cid.ipfs.tech/#zb2rhj7crUKTQYRGCRATFaQ6YFLTde2YzdqbbhAASkL9uRDXn
       expect(output == "zb2rhj7crUKTQYRGCRATFaQ6YFLTde2YzdqbbhAASkL9uRDXn", true);
@@ -18,7 +18,7 @@ void main() {
 
     test('creating cid with \'hello world \' with base32 encoding', () {
       String input = 'hello world';
-      final output = createCid(input, BASE.base32);
+      final output = Cid.createCid(input, BASE.base32);
 
       // https://cid.ipfs.tech/#BAFKREIFZJUT3TE2NHYEKKLSS27NH3K72YSCO7Y32KOAO5EEI66WOF36N5E
       expect(output == "BAFKREIFZJUT3TE2NHYEKKLSS27NH3K72YSCO7Y32KOAO5EEI66WOF36N5E", true);
@@ -27,16 +27,16 @@ void main() {
     test('different cids when input value is different', () {
       String input1 = 'divinity';
       String input2 = 'something comforting';
-      final output1 = createCid(input1, BASE.base32);
-      final output2 = createCid(input2, BASE.base32);
+      final output1 = Cid.createCid(input1, BASE.base32);
+      final output2 = Cid.createCid(input2, BASE.base32);
 
       expect(output1 == output2, false);
     });
 
     test('empty values should yield results', () {
       String input = '';
-      final output1 = createCid(input, BASE.base32);
-      final output2 = createCid(input, BASE.base58);
+      final output1 = Cid.createCid(input, BASE.base32);
+      final output2 = Cid.createCid(input, BASE.base58);
 
       expect(output1, isNotEmpty);
       expect(output1 == 'BAFKREIHDWDCEFGH4DQKJV67UZCMW7OJEE6XEDZDETOJUZJEVTENXQUVYKU', true);
@@ -103,7 +103,7 @@ Future<CidComparison> comparedPackageWithIPFSCid(String inputString, String file
 
   // Cid returned from running the command
   final ipfsCid = match?.group(1)?.trim();
-  final packageCid = createCid(inputString, BASE.base32);
+  final packageCid = Cid.createCid(inputString, BASE.base32);
 
   // Cleanup
   await process.kill();
